Divide TestTpchDistributedStats
Separate TestDisplayTpchDistributedStats as this one is very specific
test. It only displays (prints) metric comparisons. While
TestTpchDistributedStats is going to contain much more tests.
Having them merged was messy.
